+++ internal/cortex/querier/queryrange/query_range_test.go
@@ -86,15 +86,17 @@ func TestRequest(t *testing.T) {
 }
 
 func TestResponse(t *testing.T) {
-	r := *parsedResponse
-	r.Headers = respHeaders
 	for i, tc := range []struct {
 		body     string
 		expected *PrometheusResponse
 	}{
 		{
 			body:     responseBody,
-			expected: &r,
+			expected: withHeaders(parsedResponse, respHeaders),
+		},
+		{
+			body:     histogramResponseBody,
+			expected: withHeaders(parsedHistogramResponse, respHeaders),
 		},
 	} {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
@@ -666,4 +668,10 @@ func mustParse(t *testing.T, response string) Response {
 	json := jsoniter.ConfigCompatibleWithStandardLibrary
 	require.NoError(t, json.Unmarshal([]byte(response), &resp))
 	return &resp
+}
+
+func withHeaders(response *PrometheusResponse, headers []*PrometheusResponseHeader) *PrometheusResponse {
+	r := *response
+	r.Headers = headers
+	return &r
 }