+++ pkg/queryfrontend/queryinstant_codec_test.go
@@ -292,7 +292,8 @@ func TestMergeResponse(t *testing.T) {
 								Vector: &queryrange.Vector{
 									Samples: []*queryrange.Sample{
 										{
-											Sample: cortexpb.Sample{TimestampMs: 0, Value: 1},
+											Timestamp:   0,
+											SampleValue: 1,
 											Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 												"__name__": "up",
 											})),
@@ -313,7 +314,8 @@ func TestMergeResponse(t *testing.T) {
 							Vector: &queryrange.Vector{
 								Samples: []*queryrange.Sample{
 									{
-										Sample: cortexpb.Sample{TimestampMs: 0, Value: 1},
+										Timestamp:   0,
+										SampleValue: 1,
 										Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 											"__name__": "up",
 										})),
@@ -337,7 +339,8 @@ func TestMergeResponse(t *testing.T) {
 								Vector: &queryrange.Vector{
 									Samples: []*queryrange.Sample{
 										{
-											Sample: cortexpb.Sample{TimestampMs: 0, Value: 1},
+											Timestamp:   0,
+											SampleValue: 1,
 											Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 												"__name__": "up",
 												"job":      "foo",
@@ -358,7 +361,8 @@ func TestMergeResponse(t *testing.T) {
 								Vector: &queryrange.Vector{
 									Samples: []*queryrange.Sample{
 										{
-											Sample: cortexpb.Sample{TimestampMs: 0, Value: 2},
+											Timestamp:   0,
+											SampleValue: 2,
 											Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 												"__name__": "up",
 												"job":      "bar",
@@ -380,14 +384,16 @@ func TestMergeResponse(t *testing.T) {
 							Vector: &queryrange.Vector{
 								Samples: []*queryrange.Sample{
 									{
-										Sample: cortexpb.Sample{TimestampMs: 0, Value: 2},
+										Timestamp:   0,
+										SampleValue: 2,
 										Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 											"__name__": "up",
 											"job":      "bar",
 										})),
 									},
 									{
-										Sample: cortexpb.Sample{TimestampMs: 0, Value: 1},
+										Timestamp:   0,
+										SampleValue: 1,
 										Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 											"__name__": "up",
 											"job":      "foo",
@@ -412,7 +418,8 @@ func TestMergeResponse(t *testing.T) {
 								Vector: &queryrange.Vector{
 									Samples: []*queryrange.Sample{
 										{
-											Sample: cortexpb.Sample{TimestampMs: 0, Value: 1},
+											Timestamp:   0,
+											SampleValue: 1,
 											Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 												"__name__": "up",
 												"job":      "foo",
@@ -433,7 +440,8 @@ func TestMergeResponse(t *testing.T) {
 								Vector: &queryrange.Vector{
 									Samples: []*queryrange.Sample{
 										{
-											Sample: cortexpb.Sample{TimestampMs: 1, Value: 2},
+											Timestamp:   1,
+											SampleValue: 2,
 											Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 												"__name__": "up",
 												"job":      "foo",
@@ -455,7 +463,8 @@ func TestMergeResponse(t *testing.T) {
 							Vector: &queryrange.Vector{
 								Samples: []*queryrange.Sample{
 									{
-										Sample: cortexpb.Sample{TimestampMs: 1, Value: 2},
+										Timestamp:   1,
+										SampleValue: 2,
 										Labels: cortexpb.FromLabelsToLabelAdapters(labels.FromMap(map[string]string{
 											"__name__": "up",
 											"job":      "foo",
@@ -774,7 +783,8 @@ func TestDecodeResponse(t *testing.T) {
 											"instance": "localhost:9090",
 											"job":      "prometheus",
 										})),
-										Sample: cortexpb.Sample{TimestampMs: 1661020672043, Value: 1},
+										Timestamp:   1661020672043,
+										SampleValue: 1,
 									},
 								},
 							},